3.2.0 next major updates several fixes and new sub systems

+++ bin/vm/install-profiler.sh
@@ -8,37 +8,50 @@ then
 fi
 touch /home/vagrant/.profiler
 
+if [ -f /home/vagrant/.php56 ]
+then
 sudo apt-get install php5.6-xhprof
+fi
 
+if [ -f /home/vagrant/.php72 ] || [ -f /home/vagrant/.php71 ] || [ -f /home/vagrant/.php70 ]
+then
+    sudo apt-get install libcurl4-openssl-dev libpcre3-dev -y
+    cd /usr/src
+    sudo git clone https://github.com/tideways/php-profiler-extension.git
+    cd php-profiler-extension
+fi
 
-sudo apt-get install libcurl4-openssl-dev libpcre3-dev -y
-cd /usr/src
-sudo git clone https://github.com/tideways/php-profiler-extension.git
-cd php-profiler-extension
+if [ -f /home/vagrant/.php72 ]
+then
+    sudo make clean
+    sudo phpize7.2
+    sudo ./configure
+    sudo make
+    sudo make install
+    sudo ln -s /vagrant/etc/php/profiler.7.2.ini /etc/php/7.2/fpm/conf.d/20-profiler.ini
+    sudo ln -s /vagrant/etc/php/profiler.7.2.ini /etc/php/7.2/cli/conf.d/20-profiler.ini
+fi
 
+if [ -f /home/vagrant/.php71 ]
+then
+    sudo make clean
+    sudo phpize7.1
+    sudo ./configure
+    sudo make
+    sudo make install
+    sudo ln -s /vagrant/etc/php/profiler.7.1.ini /etc/php/7.1/fpm/conf.d/20-profiler.ini
+    sudo ln -s /vagrant/etc/php/profiler.7.1.ini /etc/php/7.1/cli/conf.d/20-profiler.ini
+fi
 
-sudo make clean
-sudo phpize7.2
-sudo ./configure
-sudo make
-sudo make install
-sudo ln -s /vagrant/etc/profiler.7.0.ini /etc/php/7.2/fpm/conf.d/20-profiler.ini
-sudo ln -s /vagrant/etc/profiler.7.0.ini /etc/php/7.2/cli/conf.d/20-profiler.ini
-
-sudo make clean
-sudo phpize7.1
-sudo ./configure
-sudo make
-sudo make install
-sudo ln -s /vagrant/etc/profiler.7.1.ini /etc/php/7.1/fpm/conf.d/20-profiler.ini
-sudo ln -s /vagrant/etc/profiler.7.1.ini /etc/php/7.1/cli/conf.d/20-profiler.ini
-
-sudo make clean
-sudo phpize7.0
-sudo ./configure
-sudo make
-sudo make install
-sudo ln -s /vagrant/etc/profiler.7.2.ini /etc/php/7.0/fpm/conf.d/20-profiler.ini
-sudo ln -s /vagrant/etc/profiler.7.2.ini /etc/php/7.0/cli/conf.d/20-profiler.ini
+if [ -f /home/vagrant/.php70 ]
+then
+    sudo make clean
+    sudo phpize7.0
+    sudo ./configure
+    sudo make
+    sudo make install
+    sudo ln -s /vagrant/etc/php/profiler.7.0.ini /etc/php/7.0/fpm/conf.d/20-profiler.ini
+    sudo ln -s /vagrant/etc/php/profiler.7.0.ini /etc/php/7.0/cli/conf.d/20-profiler.ini
+fi
